#INS-3586 moved mutable.go outgoing checker simplified
